requirements.txt
asgiref==3.8.1
certifi==2024.6.2
Django==5.0.6
django-cors-headers==4.3.1
django-elasticsearch-dsl==8.0
django-environ==0.11.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
elastic-transport==8.13.1
elasticsearch==8.13.2
elasticsearch-dsl==8.13.1
inflection==0.5.1
Markdown==3.6
packaging==24.0
pillow==10.3.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
six==1.16.0
sqlparse==0.5.0
uritemplate==4.1.1
urllib3==2.2.1
